refactors kan_layer errors into their own module and file
